requirements.txt
Flask==3.1.2
google-generativeai==0.8.6
google-api-python-client==2.188.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
beautifulsoup4==4.14.3
requests==2.32.5
feedparser==6.0.12
fpdf2==2.8.5
python-dotenv==1.2.1
gunicorn==21.2.0
